requirements.txt
flask
flask-restful
-e git+https://github.com/mark-burnett/jot.git#egg=jot
oauthlib
pycrypto
simplejson
simplepam
sqlalchemy
